# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is a sample go-api.\nIt also does this.",
        "title": "My API",
        "contact": {
            "name": "Jonny Langefeld",
            "email": "jonny.langefeld@gmail.com"
        },
        "license": {}
    },
    "host": "example.com",
    "basePath": "/",
    "paths": {
        "/articles": {
            "get": {
                "description": "Get all articles stored in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "List all articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the page to be retrieved",
                        "name": "page_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ArticleList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "PutArticle writes an article to the database\nTo write a new article, leave the id empty. To update an existing one, use the id of the article to be updated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Add an article to the database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "description": "GetArticle returns a single article by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "article id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "description": "Get all orders stored in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the page to be retrieved",
                        "name": "page_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OrderList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "PutOrder writes an order to the database\nTo write a new order, leave the id empty. To update an existing one, use the id of the order to be updated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Add an order to the database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "GetOrder returns a single order by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Article": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The unique id of this item",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "The name of this item",
                    "type": "string",
                    "example": "Skittles"
                },
                "price": {
                    "description": "The price of this item",
                    "type": "number",
                    "example": 1.99
                }
            }
        },
        "ArticleList": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "A list of articles",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Article"
                    }
                },
                "next_page_id": {
                    "description": "The id to query the next page",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string",
                    "example": "The requested resource was not found on the server"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string",
                    "example": "Resource not found."
                }
            }
        },
        "Order": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The unique id of this order",
                    "type": "integer",
                    "example": 1
                },
                "lastUpdated": {
                    "description": "DateTime is the date and time of this order",
                    "type": "string",
                    "example": "0001-01-01 00:00:00+00"
                }
            }
        },
        "OrderList": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "A list of orders",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Order"
                    }
                },
                "next_page_id": {
                    "description": "The id to query the next page",
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}